Raise PHPStan to level 7 (fix the ~25 findings) (#133)
Bumps static analysis from level 6 to 7 — stricter argument and return
typing — and fixes everything it surfaced, no behaviour change:

- list-returning methods that feed JSON/output now array_values() to keep
  their list<> contract provable (Role/User repos, RelationRepository,
  SchemaToolset::currentFieldDefs);
- guard/cast the false-returning builtins (Env file(), session_name(),
  strtotime(), getenv() in the test bootstrap);
- precise param types where list<> was expected (TokenPrincipal scopes,
  ContentToolset fields) and a broadened entity() key type;
- small test-helper type fixes (preg_match offsets, ReflectionFunction via
  Closure::fromCallable, query-string keys stringified).

Also marks the stale "PHP-CS-Fixer in CI" roadmap item done (it already runs
in ci.yml). 673 tests green; php-cs-fixer clean.

+++ src/Admin/EntriesController.php
@@ -291,7 +291,7 @@ final class EntriesController extends Controller
     /** A stored "Y-m-d H:i:s" as the "Y-m-d\TH:i" a datetime-local input expects. */
     private function toDatetimeLocal(?string $stored): string
     {
-        return $stored !== null && $stored !== '' ? date('Y-m-d\TH:i', strtotime($stored)) : '';
+        return $stored !== null && $stored !== '' ? date('Y-m-d\TH:i', (int) strtotime($stored)) : '';
     }
 
     /** Build the typed input object from the request (with normalized values). */
